CORE-7040: Rename module `components:virtual-node:virtual-node-info-read-service-rpc-extensions`
and associated package.
