+++ components/membership/membership-rest-impl/build.gradle
@@ -18,7 +18,7 @@ dependencies {
     implementation project(":components:membership:membership-group-read")
     implementation project(":components:membership:membership-rest")
     implementation project(":components:virtual-node:virtual-node-info-read-service")
-    implementation project(":components:virtual-node:virtual-node-info-read-service-rpc-extensions")
+    implementation project(":components:virtual-node:virtual-node-info-read-service-rest-extensions")
     implementation project(':components:configuration:configuration-read-service')
     implementation project(":libs:crypto:crypto-core")
     implementation project(":libs:messaging:messaging")
