+++ components/membership/membership-rest-impl/src/main/kotlin/net/corda/membership/impl/rest/v1/KeysRestResourceImpl.kt
@@ -29,7 +29,7 @@ import net.corda.membership.rest.v1.types.response.KeyMetaData
 import net.corda.membership.rest.v1.types.response.KeyPairIdentifier
 import net.corda.membership.impl.rest.v1.lifecycle.RestResourceLifecycleHandler
 import net.corda.v5.crypto.publicKeyId
-import net.corda.virtualnode.read.rpc.extensions.createKeyIdOrHttpThrow
+import net.corda.virtualnode.read.rest.extensions.createKeyIdOrHttpThrow
 import org.osgi.service.component.annotations.Activate
 import org.osgi.service.component.annotations.Component
 import org.osgi.service.component.annotations.Reference
