+++ components/virtual-node/virtual-node-info-read-service-rest-extensions/src/main/java/net/corda/virtualnode/read/rest/extensions/package-info.java
@@ -1,0 +1,4 @@
+@Export
+package net.corda.virtualnode.read.rest.extensions;
+
+import org.osgi.annotation.bundle.Export;
