+++ components/virtual-node/virtual-node-info-read-service-rest-extensions/src/main/kotlin/net/corda/virtualnode/read/rest/extensions/VirtualNodeInfoRPCExtensions.kt
@@ -1,4 +1,4 @@
-package net.corda.virtualnode.read.rpc.extensions
+package net.corda.virtualnode.read.rest.extensions
 
 import net.corda.crypto.core.ShortHash
 import net.corda.crypto.core.ShortHash.Companion.LENGTH
