--- components/virtual-node/virtual-node-info-read-service-rpc-extensions/src/main/java/net/corda/virtualnode/read/rpc/extensions/package-info.java
@@ -1,4 +1,0 @@
-@Export
-package net.corda.virtualnode.read.rpc.extensions;
-
-import org.osgi.annotation.bundle.Export;
