+++ components/virtual-node/virtual-node-rpcops-service-impl/build.gradle
@@ -14,7 +14,7 @@ dependencies {
     implementation project(':components:virtual-node:cpi-info-read-service')
     implementation project(':components:virtual-node:virtual-node-management-sender')
     implementation project(':components:virtual-node:virtual-node-info-read-service')
-    implementation project(":components:virtual-node:virtual-node-info-read-service-rpc-extensions")
+    implementation project(":components:virtual-node:virtual-node-info-read-service-rest-extensions")
     implementation project(':libs:crypto:crypto-core')
     implementation project(':libs:virtual-node:virtual-node-common')
     implementation project(':libs:virtual-node:virtual-node-endpoints')
