+++ settings.gradle
@@ -190,7 +190,7 @@ include 'components:virtual-node:sandbox-json'
 include 'components:virtual-node:sandbox-notary'
 include 'components:virtual-node:sandbox-notary-impl'
 include 'components:virtual-node:virtual-node-info-read-service'
-include 'components:virtual-node:virtual-node-info-read-service-rpc-extensions'
+include 'components:virtual-node:virtual-node-info-read-service-rest-extensions'
 include 'components:virtual-node:virtual-node-management-sender'
 include 'components:virtual-node:virtual-node-rpcops-service-impl'
 include 'components:virtual-node:virtual-node-rpcops-maintenance-impl'
